% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calendar.R
\docType{data}
\name{calendar}
\alias{calendar}
\title{Calendar for the cycle argument}
\format{
A data frame with 73049 rows and 6 variables:
\describe{
\item{D}{daily}
\item{SM}{semi-monthly}
\item{M}{monthly}
\item{Q}{quarterly}
\item{S}{semi-annually}
\item{A}{annually}
}
}
\usage{
calendar
}
\description{
The ecos open API has been revised on 2022.06.01.
A calendar was created to respond to any type of date format according to the cycle argument.
}
\examples{
calendar
}
\keyword{datasets}
