% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecos-package.R
\docType{package}
\name{ecos-package}
\alias{ecos}
\alias{ecos-package}
\title{ecos: Economic Statistics System of the Bank of Korea}
\description{
API wrapper to download statistical information from the Economic Statistics System (ECOS) of the Bank of Korea \url{https://ecos.bok.or.kr/}.
}
\details{
To use this package, you will first need to get your API key from the
website \url{https://ecos.bok.or.kr/api/#/AuthKeyApply}. Once you have your
key, you can save it as an environment variable for the current session
using the \code{\link{setKey}} function. Alternatively, you can set it
permanently by adding the following line to your .Renviron file:

ECOS_API_KEY = PASTE YOUR API KEY

Any functions that require your API key try to retrieve it via
\code{Sys.getenv("ECOS_API_KEY")} (unless API key is explicitly specified as
a function argument).
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

Other contributors:
\itemize{
  \item Jaehyun Joo [contributor]
}

}
