% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyStatisticList.R
\name{keyStatisticList}
\alias{keyStatisticList}
\title{keyStatisticList Function}
\usage{
keyStatisticList(api_key, format, lang, count)
}
\arguments{
\item{api_key, format, lang, count}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# 100 Economic Statistics in Korea
df <- keyStatisticList()
head(df)
}
\keyword{ecos,}
\keyword{keyStatisticList}
