% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyStatList.R
\name{keyStatList}
\alias{keyStatList}
\title{keyStatList Function}
\usage{
keyStatList(api_key, format, lang, count)
}
\arguments{
\item{api_key, format, lang, count}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# 100 Key Economic Statistics in Korea
df <- keyStatList()
head(df)
}
\keyword{ecos,}
\keyword{keyStatList}
