% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orderStatSearchColumns}
\alias{orderStatSearchColumns}
\title{orderStatSearchColumns Function}
\usage{
orderStatSearchColumns(x)
}
\arguments{
\item{x}{return object of function statSearch}
}
\description{
check and sort column order
}
\keyword{internal}
