% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setCalendar}
\alias{setCalendar}
\title{setCalendar Function}
\usage{
setCalendar(start = "1900-01-01", end = "2099-12-31")
}
\arguments{
\item{start}{Start date}

\item{end}{End date}
}
\description{
Set calendar data frame to covert date format according to a
  cycle argument
}
\keyword{internal}
