% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKey.R
\name{setKey}
\alias{setKey}
\alias{printKey}
\title{Set ECOS API Key}
\usage{
setKey(api_key)

printKey()
}
\arguments{
\item{api_key}{A string specifying ECOS API key}
}
\value{
No return value, called to set api key
}
\description{
Save ECOS API key for the current session. To set it permanently, please add
the following line to your .Renvrion file:

ECOS_API_KEY = YOUR API KEY
}
\examples{

## Set API Key for the current session
\donttest{setKey("SetYourKey")}

## Check API key
printKey()

}
