% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showStatItemList}
\alias{showStatItemList}
\title{showStatItemList Function}
\usage{
showStatItemList(stat_code, format = c("xml", "json"), lang = c("kr", "en"))
}
\arguments{
\item{stat_code}{Statistical table code}

\item{format}{File format of the result value - xml, json}

\item{lang}{Language of result value - kr (Korean), en (English)}
}
\description{
Show neat return object of statItemList function
}
\keyword{internal}
