% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticSearch.R
\name{staticSearch}
\alias{staticSearch}
\title{staticSearch Function}
\usage{
staticSearch(key, stat_code, count, cycle, start, end, item_code)
}
\arguments{
\item{you}{can get korea economy data through the api}
}
\description{
ECOS is the economy statistics (from Bank of Korea, http://ecos.bok.or.kr)
}
\examples{
df <- staticSearch(key = "sample", stat_code = "010Y002", cycle = "MM", start = "196001", end = "201812", item_code = "AAAA11")
head(df)
}
\keyword{ecos,}
\keyword{staticSearch}
