% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statItemList.R
\name{statItemList}
\alias{statItemList}
\title{statItemList Function}
\usage{
statItemList(
  api_key,
  format = c("xml", "json"),
  lang = c("kr", "en"),
  count,
  stat_code
)
}
\arguments{
\item{api_key}{Open API authentication key issued by the Bank of Korea}

\item{format}{File format of the result value - xml, json}

\item{lang}{Language of result value - kr (Korean), en (English)}

\item{count}{Number of requests}

\item{stat_code}{Statistical table code}
}
\description{
Item List of Statistics
}
\examples{
# do not test -- needs an API key
# statItemList(api_key = your_api_key, lang = "en", count = 100, stat_code = "902Y001")
# statItemList(api_key = your_api_key, lang = "kr", count = 100, stat_code = "902Y001")

}
\keyword{ecos,}
\keyword{statItemList}
