% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{statMeta}
\alias{statMeta}
\title{Statistics Meta}
\usage{
statMeta(api_key, format = c("xml", "json"), lang = c("kr", "en"), count, meta)
}
\arguments{
\item{api_key, format, lang, count, meta}{input parameters}
}
\description{
statMeta Function
}
\examples{
statMeta(api_key = your_api_key, format = "xml", lang = "en", count = 16, meta = "Economic Sentiment Index")
statMeta(api_key = your_api_key, format = "xml", lang = "kr", count = 16, meta = "경제심리지수")

}
\keyword{ecos,}
\keyword{statMeta}
