% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statSearch.R
\name{statSearch}
\alias{statSearch}
\title{statSearch Function}
\usage{
statSearch(api_key, format, lang, count, stat_code, cycle, start_date,
  end_date, item_code)
}
\arguments{
\item{api_key, format, lang, count, stat_code, cycle, start_date, end_date, item_code}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# stock index futures 
df <- statSearch(api_key = your_api_key, format = "xml", lang = "kr", count = 1000, stat_code = "085Y007", cycle = "MM", start_date = "196001", end_date = "201812", item_code = "S25B")
head(df)
}
\keyword{ecos,}
\keyword{statSearch}
