% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTableList.R
\name{statTableList}
\alias{statTableList}
\title{statTableList Function}
\usage{
statTableList(api_key, format, lang, count)
}
\arguments{
\item{api_key, format, lang, count}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# Statistics Table List 
df <- statTableList()
head(df)
}
\keyword{ecos,}
\keyword{statTableList}
