% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statWord.R
\name{statWord}
\alias{statWord}
\title{statWord Function}
\usage{
statWord(api_key, format = c("xml", "json"), lang = c("kr", "en"), count, word)
}
\arguments{
\item{api_key}{Open API authentication key issued by the Bank of Korea}

\item{format}{File format of the result value - xml, json}

\item{lang}{Language of result value - kr (Korean), en (English)}

\item{count}{Number of requests}

\item{word}{Terms}
}
\description{
Glossary of Statistical Terms
}
\examples{
# do not test -- needs an API key
# statWord(api_key = your_api_key, lang = "kr", count = 10, word = "CPIS")
# statWord(api_key = your_api_key, lang = "kr", count = 10, word = "소비자동향지수")

}
\keyword{ecos,}
\keyword{statMeta}
