% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyStatList.R
\name{keyStatList}
\alias{keyStatList}
\title{Top 100 Statistical Indicators}
\usage{
keyStatList(api_key, format = c("xml", "json"), lang = c("kr", "en"), count)
}
\arguments{
\item{api_key, format, lang, count}{input parameters}
}
\description{
keyStatList Function
}
\examples{
# do not test -- needs an API key
# keyStatList(api_key = your_api_key, lang = "en", count = 100)
# keyStatList(api_key = your_api_key, lang = "kr", count = 100)

}
\keyword{ecos,}
\keyword{keyStatList}
