% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyStatList.R
\name{keyStatList}
\alias{keyStatList}
\title{Top 100 statistical indicators}
\usage{
keyStatList(format = c("xml", "json"), lang = c("kr", "en"), count = 1000)
}
\arguments{
\item{format}{A string specifying the file format to process - xml, json}

\item{lang}{A string specifying the language of result value - kr (Korean),
en (English)}

\item{count}{An integer specifying the number of requests}
}
\value{
A data.frame object containing queried information
}
\description{
Top 100 statistical indicators
}
\details{
\preformatted{
## Example
keyStatList(lang = "en", count = 100)
}
}
