% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showStatItemList}
\alias{showStatItemList}
\title{showStatItemList Function}
\usage{
showStatItemList(
  api_key,
  format = c("xml", "json"),
  lang = c("kr", "en"),
  stat_code
)
}
\arguments{
\item{api_key}{Open API authentication key issued by the Bank of Korea}

\item{format}{File format of the result value - xml, json}

\item{lang}{Language of result value - kr (Korean), en (English)}

\item{stat_code}{Statistical table code}
}
\description{
Show neat return object of statItemList function
}
\keyword{internal}
