% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showStatTableList}
\alias{showStatTableList}
\title{showStatTableList Function}
\usage{
showStatTableList(api_key, format = c("xml", "json"), lang = c("kr", "en"))
}
\arguments{
\item{api_key}{Open API authentication key issued by the Bank of Korea}

\item{format}{File format of the result value - xml, json}

\item{lang}{Language of result value - kr (Korean), en (English)}
}
\description{
Show neat return object of statTableList function
}
\keyword{internal}
