% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticSearch.R
\name{staticSearch}
\alias{staticSearch}
\title{staticSearch Function}
\usage{
staticSearch(key, format, lang, count, stat_code, cycle, start_date,
  end_date, item_code)
}
\arguments{
\item{you}{can get korea economy data through the api}
}
\description{
ECOS is the economy statistics (from Bank of Korea, http://ecos.bok.or.kr)
}
\examples{
df <- staticSearch()
head(df)
}
\keyword{ecos,}
\keyword{staticSearch}
