% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticSearch.R
\name{staticSearch}
\alias{staticSearch}
\title{staticSearch Function}
\usage{
staticSearch(api_key, format, lang, count, stat_code, cycle, start_date,
  end_date, item_code)
}
\arguments{
\item{api_key, }{format, lang, count, stat_code, cycle, start_date, end_date, item_code input parameter}
}
\description{
You can access economic statistics from Bank of Korea through the open API (http://ecos.bok.or.kr)
}
\examples{
# basic
df <- staticSearch()
head(df)

# stock index futures 
df <- staticSearch(stat_code = "085Y007", item_code = "S25B")
head(df)
}
\keyword{ecos,}
\keyword{staticSearch}
