% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statItemList.R
\name{statItemList}
\alias{statItemList}
\title{Item List of Statistics}
\usage{
statItemList(
  api_key,
  format = c("xml", "json"),
  lang = c("kr", "en"),
  count,
  stat_code
)
}
\arguments{
\item{api_key, format, lang, count, stat_code}{input parameters}
}
\description{
statItemList Function
}
\examples{
statItemList(api_key = your_api_key, format = "xml", lang = "en", count = 100, stat_code = "010Y002")
statItemList(api_key = your_api_key, format = "xml", lang = "kr", count = 100, stat_code = "010Y002")

}
\keyword{ecos,}
\keyword{statItemList}
