% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{statMeta}
\alias{statMeta}
\title{statMeta Function}
\usage{
statMeta(api_key, format = c("xml", "json"), lang = c("kr", "en"), count, meta)
}
\arguments{
\item{api_key}{Open API authentication key issued by the Bank of Korea}

\item{format}{File format of the result value - xml, json}

\item{lang}{Language of result value - kr (Korean), en (English)}

\item{count}{Number of requests}

\item{meta}{data name}
}
\description{
Statistics Meta DB
}
\examples{
# do not test -- needs an API key
# statMeta(api_key = your_api_key, lang = "en", meta = "Economic Sentiment Index")
# statMeta(api_key = your_api_key, lang = "kr", meta = "경제심리지수")

}
\keyword{ecos,}
\keyword{statMeta}
