% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{statMeta}
\alias{statMeta}
\title{statMeta Function}
\usage{
statMeta(api_key, format, lang, count, data_name)
}
\arguments{
\item{api_key, format, lang, count, data_name}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# Economic Statistics Meta in Korea
df <- statMeta()
head(df)
}
\keyword{ecos,}
\keyword{statMeta}
