% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{statMeta}
\alias{statMeta}
\title{Retrieve statistical meta DB}
\usage{
statMeta(meta, format = c("xml", "json"), lang = c("kr", "en"), count = 1000)
}
\arguments{
\item{meta}{A string specifying the name of meta DB to query}

\item{format}{A string specifying the file format to process - xml, json}

\item{lang}{A string specifying the language of result value - kr (Korean),
en (English)}

\item{count}{An integer specifying the number of requests}
}
\value{
A data.frame object containing queried information
}
\description{
Retrieve statistical meta DB
}
\details{
\preformatted{
## Example
statMeta(lang = "en", meta = "Economic Sentiment Index")
}
}
