% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statSearch.R
\name{statSearch}
\alias{statSearch}
\title{statSearch Function}
\usage{
statSearch(api_key, format, lang, count, stat_code, cycle, start_date,
  end_date, item_code)
}
\arguments{
\item{api_key, format, lang, count, stat_code, cycle, start_date, end_date, item_code}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# basic
df <- statSearch()
head(df)

# stock index futures 
df <- statSearch(stat_code = "085Y007", item_code = "S25B")
head(df)
}
\keyword{ecos,}
\keyword{statSearch}
