% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statSearch.R
\name{statSearch}
\alias{statSearch}
\title{statSearch Function}
\usage{
statSearch(
  api_key,
  format = c("xml", "json"),
  lang = c("kr", "en"),
  stat_code,
  item_code,
  item_code2,
  item_code3,
  cycle,
  start_time,
  end_time,
  count
)
}
\arguments{
\item{api_key, format, lang, count, stat_code, cycle, start_time, end_time, item_code, item_code2, item_code3}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# stock index futures 
df <- statSearch(api_key = your_api_key, format = "xml", lang = "kr", stat_code = "085Y007", item_code = "S25B", cycle = "MM", start_time = "196001", end_time = "201812", count = 1000)
head(df)
}
\keyword{ecos,}
\keyword{statSearch}
