% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTableList.R
\name{statTableList}
\alias{statTableList}
\title{List of Statistical Tables}
\usage{
statTableList(api_key, format = c("xml", "json"), lang = c("kr", "en"), count)
}
\arguments{
\item{api_key, format, lang, count}{input parameters}
}
\description{
statTableList Function
}
\examples{
statTableList(api_key = your_api_key, lang = "en", count = 1000)
statTableList(api_key = your_api_key, lang = "kr", count = 1000)

}
\keyword{ecos,}
\keyword{statTableList}
