% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTableList.R
\name{statTableList}
\alias{statTableList}
\title{statTableList Function}
\usage{
statTableList(api_key, format, lang, count)
}
\arguments{
\item{api_key, format, lang, count}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the OPEN API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# Economic Statistics Table List from Bank of Korea through the OPEN API
df <- statTableList(api_key = your_api_key, format = "xml", lang = "kr", count = 1000)
head(df)
}
\keyword{ecos,}
\keyword{statTableList}
