% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statWord.R
\name{statWord}
\alias{statWord}
\title{statWord Function}
\usage{
statWord(api_key, format, lang, count, word)
}
\arguments{
\item{api_key, format, lang, count, word}{input parameters}
}
\description{
You can access economic statistics from Bank of Korea through the open API (https://ecos.bok.or.kr/jsp/openapi/OpenApiController.jsp)
}
\examples{
# Economic Statistics word from Bank of Korea through the OPEN API
df <- statWord()
head(df)
}
\keyword{ecos,}
\keyword{statMeta}
