% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statWord.R
\name{statWord}
\alias{statWord}
\title{Glossary of Statistical Terms}
\usage{
statWord(api_key, format = c("xml", "json"), lang = c("kr", "en"), count, word)
}
\arguments{
\item{api_key, format, lang, count, word}{input parameters}
}
\description{
statWord Function
}
\examples{
statWord(api_key = your_api_key, lang = "kr", count = 10, word = "CPIS")
statWord(api_key = your_api_key, lang = "kr", count = 10, word = "소비자동향지수")

}
\keyword{ecos,}
\keyword{statMeta}
