% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statWord.R
\name{statWord}
\alias{statWord}
\title{Glossary of Statistical Terms}
\usage{
statWord(
  api_key,
  word,
  format = c("xml", "json"),
  lang = c("kr", "en"),
  count = 1000
)
}
\arguments{
\item{api_key}{A string specifying ECOS API key. Need not be specified if
the key was stored as an environment variable via \code{\link{setKey}} or
.Renviron.}

\item{word}{A string specifying the term to search}

\item{format}{A string specifying the file format to process - xml, json}

\item{lang}{A string specifying the language of result value - kr (Korean),
en (English)}

\item{count}{An integer specifying the number of requests}
}
\value{
A data.frame object containing queried information
}
\description{
Glossary of Statistical Terms
}
\details{
\preformatted{
## Example
statWord(word = "CPI", lang = "en")
}
}
